Update to upstream snapshot for GNOME 45 compatibility
I'm bumping the snapshot number from 11 to 20 in this change.  That is
because this snapshot is from an upstream branch that only works on
GNOME 45.  If at some point I need to update other dist-git branches
that don't have GNOME 45, I need to be able to use snapshots of other
upstream branches and still have a correct upgrade path.  I'm leaving a
relatively small window (eight digits, eight snapshots) because I don't
think this is likely, but I don't want to rule it out either.